Added .keep-account file logic
A users processes are all killed after 2 weeks of not being logged in. If this is not desired a user should be able to signal they want their processes to stay alive. This change should let a user perform "touch ~/.keep-account" as a signal to prune.sh to skip them during pruning.